LayoutTests/fast/mediastream/MediaDevices-getUserMedia-expected.txt

Tests getUserMedia (Promise-based version on navigator.mediaDevices)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof navigator.mediaDevices.webkitGetUserMedia is 'undefined'
PASS navigator.mediaDevices.getUserMedia({audio:true}).then(gotStream1); did not throw exception.
PASS navigator.mediaDevices.getUserMedia() rejected with error: TypeError: Not enough arguments
PASS  navigator.mediaDevices.getUserMedia({}) rejected with error: NotSupportedError (DOM Exception 9): The operation is not supported.
PASS Stream generated.
PASS stream.getAudioTracks().length is 1
PASS stream.getVideoTracks().length is 0
PASS navigator.mediaDevices.getUserMedia({video:true}).then(gotStream2); did not throw exception.
PASS Stream generated.
PASS stream.getAudioTracks().length is 0
PASS stream.getVideoTracks().length is 1
PASS navigator.mediaDevices.getUserMedia({audio:true, video:true}).then(gotStream3); did not throw exception.
PASS Stream generated.
PASS stream.getAudioTracks().length is 1
PASS stream.getVideoTracks().length is 1
PASS navigator.mediaDevices.getUserMedia({audio:{}, video:{}}).then(gotStream4); did not throw exception.
PASS Stream generated.
PASS stream.getAudioTracks().length is 1
PASS stream.getVideoTracks().length is 1
PASS navigator.mediaDevices.getUserMedia({audio:true}).then(invalidGotStream, error1); did not throw exception.
PASS Error callback called.
PASS errorArg.name is "NotAllowedError"
PASS navigator.mediaDevices.getUserMedia({audio:true}).then(invalidGotStream).catch(error2); did not throw exception.
PASS Error callback called.
PASS successfullyParsed is true

TEST COMPLETE
